✨ Refactor overlay service imports and update related functions for consistency
