+++ src/app/api/overlay/route.ts
@@ -1,10 +1,10 @@
-import { overlayService } from '@/db';
+import { OverlayService } from '@/lib/overlayService';
 
 export async function POST(request: Request) {
   try {
     const config = await request.json();
 
-    const overlayId = await overlayService.createOverlay(config);
+    const overlayId = await OverlayService.createOverlay(config);
 
     return Response.json({ overlayId });
   } catch (error) {
