+++ src/lib/db.ts
@@ -1,0 +1,19 @@
+import Database from 'better-sqlite3';
+import { drizzle } from 'drizzle-orm/better-sqlite3';
+import fs from 'fs';
+import path from 'path';
+
+const dbPath = path.join(process.cwd(), '.overlay-storage', 'overlays.db');
+
+// Ensure directory exists
+const dir = path.dirname(dbPath);
+if (!fs.existsSync(dir)) {
+  fs.mkdirSync(dir, { recursive: true });
+}
+
+const sqlite = new Database(dbPath);
+sqlite.pragma('journal_mode = WAL');
+sqlite.pragma('wal_autocheckpoint = 1000');
+sqlite.pragma('cache_size = -2000');
+
+export const db = drizzle(sqlite);
